feat(users): Implement More LFA feedbacks
- Write more codes to test different scenarios of user input and request
- Ensure that data type entered by user on endpoints that require parameters are numbers only
- trim white spaces during validation
- Change userinput for email to lowecase on login request
- Change userinput for email and secondaryEmail to lowercase on signup request
- create more codes for more instances of user error on controllers

[Delivers #164743351]

+++ src/v2/controllers/groupsControllers.js
@@ -1,6 +1,4 @@
-import passport from 'passport';
-import passportConf from "../passport";
-import db from "../db/index";
+import db from '../db/index';
 
 const Group = {
   async createGroup(req, res) {
@@ -11,16 +9,16 @@ const Group = {
     const values = [role, name, user.email];
     try {
       const { rows } = await db.query(text, values);
-      return res.status(201).json({status: 201, data: [rows[0]],});
-    }catch (error) {
-      return res.status(400).json({ status: 400, error });
+      return res.status(201).json({ status: 201, data: [rows[0]] });
+    } catch (error) {
+      return res.status(500).json({ status: 500, error });
     }
   },
 
   async getAllGroups(req, res) {
     const user = req.decodedToken;
     const text = 'SELECT * FROM groupTable';
-    const values = [];
+    // const values = [user[0].email];
     try {
       const { rows } = await db.query(text);
       if (!rows[0]) {
@@ -34,23 +32,30 @@ const Group = {
 
   async editGroupName(req, res) {
     const user = req.decodedToken;
-    const { id, name } = req.params;
+    const idToNum = Number(req.params.id);
+    if (isNaN(idToNum)) {
+      return res.status(400).json({ status: 400, error: 'Bad request: Please ensure that the groupId is an integer' });
+    }
+    console.log(idToNum);
+    const { name } = req.value.body;
     const text = 'SELECT * FROM groupTable WHERE id=$1';
-    const values = [id];
+    const values = [idToNum];
+    console.log('here mow');
     try {
       const { rows } = await db.query(text, values);
 
       if (!rows[0]) {
         return res.status(404).json({ status: 404, error: 'Not Found: Please check the provided groupId' });
       }
+      console.log('welcome back');
       if (rows[0].createdby !== user.email) {
-        return res.status(401).json({ status: 401, error: 'Unauthorized, You do not have the authority to rename this group' });
+        return res.status(403).json({ status: 403, error: 'Unauthorized, You do not have the authority to rename this group' });
       }
       if (rows[0].createdby === user.email) {
-        const text = 'UPDATE groupTable SET name=$1 WHERE id=$2 returning *';
-        const values = [name, id];
-        const { rows } = await db.query(text, values);
-        return res.status(200).json({ status: 200, data: [rows[0]] });
+        const updateText = 'UPDATE groupTable SET name=$1 WHERE id=$2 returning *';
+        const updateValues = [name, idToNum];
+        const { rows: updated } = await db.query(updateText, updateValues);
+        return res.status(200).json({ status: 200, data: [updated[0]] });
       }
     } catch (error) {
       return res.status(400).json({ status: 400, error });
@@ -61,39 +66,43 @@ const Group = {
   async deleteSpecificGroup(req, res) {
     const user = req.decodedToken;
     const text = 'SELECT * FROM groupTable WHERE id=$1';
-    const value = [req.params.id];
+    const idToNum = Number(req.params.id);
+    if (isNaN(idToNum)) {
+      return res.status(400).json({ status: 400, data: 'Bad request, Please ensure that the groupId supplied is an integer' });
+    }
+    const value = [idToNum];
     try {
       const { rows } = await db.query(text, value);
       if (!rows[0]) {
-        return res.status(404).json({ status: 404, error: `Not Found, There is no group with id=${req.params.id}` });
+        return res.status(404).json({ status: 404, error: `Not Found, There is no group with id=${idToNum}` });
       }
       if (rows[0].createdby !== user.email) {
-        return res.status(401).json({ status: 401, error: 'Unauthorized, You do not have the authority to delete this group' });
+        return res.status(403).json({ status: 403, error: 'Unauthorized, You do not have the authority to delete this group' });
       }
       if (rows[0].createdby === user.email) {
-        const text = 'DELETE FROM groupTable WHERE id=$1 returning *';
-        try {
-          const { rows } = await db.query(text, [req.params.id]);
-          return res.status(200).json({ status: 200, data: 'Group has been successfully deleted' });
-        } catch (error) {
-          return res.status(400).json({ status: 400, error });
-        }
+        const deleteText = 'DELETE FROM groupTable WHERE id=$1 returning *';
+        const { rows: deleteGroup } = await db.query(deleteText, [idToNum]);
+        return res.status(200).json({ status: 200, data: 'Group has been successfully deleted' });
       }
-    } catch(error) {
+    } catch (error) {
       return res.status(400).json({ status: 400, error });
     }
   },
 
   async addNewMember(req, res) {
-    const user = req.decodedToken; const 
-value = [req.params.id];
+    const user = req.decodedToken;
+    const idToNum = Number(req.params.id);
+    if (isNaN(idToNum)) {
+      return res.status(400).json({ status: 400, error: 'Bad request, please ensure that the groupId is an integer' });
+    }
+    const value = [idToNum];
     const text = 'SELECT * FROM groupTable WHERE id=$1';
 
     const { userEmailAddress, userRole } = req.value.body;
     try {
       const { rows } = await db.query(text, value);
       if (!rows[0]) {
-        return res.status(404).json({ status: 404, error: `Not Found: There is no group with id=${req.params.id}` });
+        return res.status(404).json({ status: 404, error: `Not Found: There is no group with id=${idToNum}` });
       }
       if (rows[0].createdby === user.email) {
         const userExistsText = 'SELECT * FROM usersTable WHERE email=$1';
@@ -107,7 +116,7 @@ value = [req.params.id];
 
         // Check if the user already exists in this group to prevent double occurence in the group
         const searchText = 'SELECT * FROM groupmemberstable WHERE (userid=$1 AND groupId=$2)';
-        const searchValues = [userExists[0].id, req.params.id];
+        const searchValues = [userExists[0].id, idToNum];
 
         const { rows: memberExistInGroup } = await db.query(searchText, searchValues);
 
@@ -117,22 +126,26 @@ value = [req.params.id];
         // If the email does not already exist in the group
         const addMemberText = `INSERT INTO groupMembersTable(groupId, userId, userRole)
                             VALUES($1,$2,$3) returning *`;
-        const addMemberValue = [req.params.id, userExists[0].id, userRole];
+        const addMemberValue = [idToNum, userExists[0].id, userRole];
         const { rows: addMemberToGroup } = await db.query(addMemberText, addMemberValue);
         return res.status(201).json({ status: 201, data: addMemberToGroup });
-
       }
 
-      return res.status(401).json({ status: 401, error: 'Unauthorized: You do not have the authority to add a user to this group' });
+      return res.status(403).json({ status: 403, error: 'Unauthorized: You do not have the authority to add a user to this group' });
     } catch (error) {
       return res.status(400).json({ status: 400, error });
     }
   },
 
   async deleteSpecificUserFromGroup(req, res) {
+    const groupId = Number(req.params.groupId);
+    const userId = Number(req.params.userId);
+    if (isNaN(groupId) || isNaN(userId)) {
+      return res.status(400).json({ status: 400, error: 'Please ensure the userId and groupId are numbers' });
+    }
     const user = req.decodedToken;
     const theGroupText = 'SELECT * FROM groupTable WHERE id=$1';
-    const theGroupValue = [req.params.groupId];
+    const theGroupValue = [groupId];
 
     try {
       // Search groupTable to see if the group exist, access the creator if it does
@@ -143,7 +156,7 @@ value = [req.params.id];
       // Group exists confirm and user owns the group
       if (theGroup[0].createdby === user.email) {
         const theMemberText = 'SELECT * FROM groupMembersTable WHERE groupId=$1 AND userId=$2';
-        const theMemberValue = [req.params.groupId, req.params.userId];
+        const theMemberValue = [groupId, userId];
         const { rows: theMember } = await db.query(theMemberText, theMemberValue);
         // The member to be deleted does not exist in the group
         if (!theMember[0]) {
@@ -151,11 +164,11 @@ value = [req.params.id];
         }
         // Member exists, delete the member from the group
         const deleteText = 'DELETE FROM groupMembersTable WHERE groupId=$1 AND userId=$2 returning *';
-        const deleteValues = [req.params.groupId, req.params.userId];
+        const deleteValues = [groupId, req.params.userId];
         const { rows: deletedMember } = await db.query(deleteText, deleteValues);
         return res.status(200).json({ status: 200, data: 'Member deleted from group successfully' });
       }
-      return res.status(401).json({ status: 401, error: 'Unauthorized: You do not have authority to delete users from this group' });
+      return res.status(403).json({ status: 403, error: 'Unauthorized: You do not have authority to delete users from this group' });
     } catch (error) {
       res.status(400).json({ status: 400, error });
     }
@@ -186,24 +199,22 @@ value = [req.params.id];
       // Clean repetitions in userid
       const allMembers = Array.from(new Set(userIdContainer));
       allMembers.forEach(async (member) => {
-          // Get email address of each member of the group
+        // Get email address of each member of the group
         const memberEmailText = 'SELECT * FROM usersTable WHERE id=$1';
         const { rows: membersEmail } = await db.query(memberEmailText, [member]);
         // emailContainer.push(membersEmail[0].email);;
         const sendInboxText = `INSERT INTO messagesTable(subject, message, parentMessageId, senderEmail, receiverEmail, status)
                   VALUES($1, $2, $3, $4, $5, $6) returning *`;
         const sendInboxValues = [subject, message, parentMessagedId || null, user.email, (membersEmail[0]).email, 'inbox'];
-        const {rows: eachSentEmail} = await db.query(sendInboxText, sendInboxValues);
-        
-        emailContainer.push(eachSentEmail[0]); 
+        const { rows: eachSentEmail } = await db.query(sendInboxText, sendInboxValues);
 
-        // Response is sent when the last email as been sent 
-        if(member === allMembers[allMembers.length - 1]){
-            return res.status(201).json({ status: 201, data: emailContainer });
+        emailContainer.push(eachSentEmail[0]);
+
+        // Response is sent when the last email as been sent
+        if (member === allMembers[allMembers.length - 1]) {
+          return res.status(201).json({ status: 201, data: emailContainer });
         }
       });
-      
-
     } catch (error) {
       return res.status(400).json({ error: 400, error });
     }
